requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
django-crispy-forms
crispy-bootstrap5
django-allauth
djangorestframework
requests
PyJWT
cryptography
Pillow
Psycopg2-binary
django-widget-tweaks